Added buggy initial try at build system. Still needs to be written for windows/linux, documented and tested. (It does not really work properly at the moment, even on OS X) Made minor fixes to code elsewhere. Made useless changes to main.cpp

+++ trunk/PauseInitiatorID.cpp
@@ -3,7 +3,7 @@
 
 namespace hg {
 PauseInitiatorID::PauseInitiatorID(pauseinitiatortype::PauseInitiatorType type,
-                                   unsigned int ID,
+                                   size_t ID,
                                    size_t timelineLength) :
 type_(type),
 ID_(ID),
